+++ accumulo-iterators/src/main/java/org/vertexium/accumulo/iterator/model/KeyBase.java
@@ -1,7 +1,27 @@
 package org.vertexium.accumulo.iterator.model;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public abstract class KeyBase {
     public static final char VALUE_SEPARATOR = '\u001f';
+
+    public static String[] splitOnValueSeparator(String s) {
+        List<String> results = new ArrayList<>();
+        int last = 0;
+        int i = s.indexOf(VALUE_SEPARATOR);
+        while (true) {
+            if (i > 0) {
+                results.add(s.substring(last, i));
+                last = i + 1;
+                i = s.indexOf(VALUE_SEPARATOR, last);
+            } else {
+                results.add(s.substring(last));
+                break;
+            }
+        }
+        return results.toArray(new String[results.size()]);
+    }
 
     public static String[] splitOnValueSeparator(String s, int partCount) {
         String[] results = new String[partCount];